LayoutTests/webaudio/AudioParam/audioparam-negative-exponentialRamp-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "both negative values" 
PASS Executing "negative-end" 
PASS Executing "positive-end" 
PASS Executing "propagate" 
PASS Audit report 
PASS > [both negative values]  
FAIL X Negative exponential ramp from -1 to -2 expected to be equal to the array [-1,-1.0001155138015747,-1.0002310276031494,-1.0003465414047241,-1.0004620552062988,-1.0005775690078735,-1.0006932020187378,-1.000808835029602,-1.0009244680404663,-1.0010401010513306,-1.0011557340621948,-1.001271367073059,-1.0013870000839233,-1.0015026330947876,-1.0016182661056519,-1.0017340183258057...] but differs in 6000 places:
	Index	Actual			Expected
	[0]	1.0000000000000000e+0	-1.0000000000000000e+0
	[1]	1.0000000000000000e+0	-1.0001155138015747e+0
	[2]	1.0000000000000000e+0	-1.0002310276031494e+0
	[3]	1.0000000000000000e+0	-1.0003465414047241e+0
	...and 5996 more errors. assert_true: expected true got false
FAIL < [both negative values] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [negative-end]  
FAIL X Exponential ramp from 2 to -1: Expected 2 for all values but found 6000 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 5996 more errors. assert_true: expected true got false
FAIL < [negative-end] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [positive-end]  
FAIL X Exponential ramp from -1 to 1: Expected -1 for all values but found 6000 unexpected values: 
	Index	Actual
	[0]	1
	[1]	1
	[2]	1
	[3]	1
	...and 5996 more errors. assert_true: expected true got false
FAIL < [positive-end] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [propagate]  
FAIL X Exponential ramp from -1 to 1 after the end of the linear ramp: Expected -1 for all values but found 1500 unexpected values: 
	Index	Actual
	[0]	-0.9980001449584961
	[1]	-0.9980001449584961
	[2]	-0.9980001449584961
	[3]	-0.9980001449584961
	...and 1496 more errors. assert_true: expected true got false
PASS   Exponential ramp after end of ramp contains only the constant 1. 
FAIL < [propagate] 1 out of 2 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 4 out of 4 tasks were failed. assert_true: expected true got false
